+++ src/main/java/com/simisinc/platform/presentation/controller/WebPackage.java
@@ -17,7 +17,12 @@ package com.simisinc.platform.presentation.controller;
 
 import java.io.Serializable;
 
-/** Used for dynamically adding script tags to the web page; Web packages are defined in web-packages.json */
+/**
+ * Used for dynamically adding script and stylesheet tags to the web page; Web packages are defined in web-packages.json
+ *
+ * @author matt rajkowski
+ * @created 11/3/24 8:00 AM
+ */
 public class WebPackage implements Serializable {
   static final long serialVersionUID = -8484048371911908893L;